+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageResource.java
@@ -1,7 +1,12 @@
 package com.teamscale.jacoco.agent.testimpact;
 
-import java.io.IOException;
-import java.util.List;
+import com.teamscale.client.ClusteredTestDetails;
+import com.teamscale.client.PrioritizableTestCluster;
+import com.teamscale.jacoco.agent.JacocoRuntimeController;
+import com.teamscale.jacoco.agent.ResourceBase;
+import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
+import com.teamscale.report.testwise.model.TestExecution;
+import com.teamscale.report.testwise.model.TestInfo;
 
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
@@ -12,14 +17,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-
-import com.teamscale.client.ClusteredTestDetails;
-import com.teamscale.client.PrioritizableTestCluster;
-import com.teamscale.jacoco.agent.JacocoRuntimeController;
-import com.teamscale.jacoco.agent.ResourceBase;
-import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
-import com.teamscale.report.testwise.model.TestExecution;
-import com.teamscale.report.testwise.model.TestInfo;
+import java.io.IOException;
+import java.util.List;
 
 /**
  * The resource of the Jersey + Jetty http server holding all the endpoints specific for the
@@ -57,7 +56,7 @@ public class TestwiseCoverageResource extends ResourceBase {
 			handleBadRequest("Test name is missing!");
 		}
 
-		logger.debug("Start test " + testId);
+		logger.debug("Start test {}", testId);
 
 		testwiseCoverageAgent.testEventHandler.testStart(testId);
 		return Response.noContent().build();
@@ -68,12 +67,12 @@ public class TestwiseCoverageResource extends ResourceBase {
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/test/end/{" + TEST_ID_PARAMETER + "}")
 	public TestInfo handleTestEnd(@PathParam(TEST_ID_PARAMETER) String testId,
-								  TestExecution testExecution) throws JacocoRuntimeController.DumpException, CoverageGenerationException {
+			TestExecution testExecution) throws JacocoRuntimeController.DumpException, CoverageGenerationException {
 		if (testId == null || testId.isEmpty()) {
 			handleBadRequest("Test name is missing!");
 		}
 
-		logger.debug("End test " + testId);
+		logger.debug("End test {}", testId);
 
 		return testwiseCoverageAgent.testEventHandler.testEnd(testId,
 				testExecution);